The fourth argument to setsockopt() is "int" on Unix and "char" on Windows. (dm) uint32_t is "unsigned int" on Unix and "unsigned long" on Windows. (dm)
git-svn-id: svn://mielke.cc/main/brltty@1285 91a5dbb7-01b9-0310-9b5f-b28072856b6e

+++ Programs/rangelist.c
@@ -76,7 +76,7 @@ void DisplayRangeList(rangeList *l)
   else {
     rangeList *c = l;
     while (1) {
-      printf("[%u..%u]",c->x,c->y);
+      printf("[%lu..%lu]",(unsigned long)c->x,(unsigned long)c->y);
       if (c->next==NULL) break;
       printf(",");
       c = c->next;
